"Add Hospital contract with doctor and patient structures, modifier functions and event logs"
This commit adds a new Solidity contract named "Hospital" with associated structures for doctors and patients, including mappings and arrays to store relevant information. Modifier functions are added for checking doctor and patient authorization, as well as for owner and staff permissions. Event logs are also included to track patient and doctor registrations, as well as doctor access revocation and granting.